(feat/src/client/view/payVIew/Farith): add connection with back for pay products
