+++ client/src/views/cart/Pay.css
@@ -1,6 +1,63 @@
-.pay{
+/* Pay.css */
+
+.pay-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 20px;
+    font-family: "Asap", sans-serif;
+    color: #333;
+  }
+  
+  .product-name {
+    font-size: 24px;
+    font-weight: bold;
+    margin-bottom: 10px;
+    font-family: Asap;
+  }
+  
+  .product-description {
+    padding-top: 9rem;
+    font-size: 16px;
     text-align: center;
-    margin: 10rem 10rem 10rem 10rem;
-    font-family: 'Courier New', Courier, monospace;
-    color: #517f7f;
-}
+    margin-bottom: 20px;
+  }
+  
+  .product-price {
+    font-size: 20px;
+    color: #e91e63;
+    margin-bottom: 10px;
+  }
+  
+  .pay-button {
+    background-color: #c67d96bd;
+    height: 2.5rem;
+    color: #fff;
+    border: none;
+    padding: 10px 20px;
+    border-radius: 0.25rem;
+    font-size: 0.8rem;
+    cursor: pointer;
+    text-align: center;
+    transition: background-color 0.3s, transform 0.3s;
+  }
+  
+  .pay-button:hover {
+    background-color: #d81b5e;
+    transform: scale(1.05);
+  }
+  
+  .cart-link {
+    margin-top: 20px;
+    color: #e91e63a1;
+    text-decoration: none;
+    font-weight: bold;
+    font-size: 13px;
+    transition: color 0.3s;
+    padding-bottom: 8rem;
+  }
+  
+  .home-link:hover {
+    color: #d81b5e;
+  }
+
