--- client/src/views/mercadoPago/comprar/mercadoPago.css
@@ -1,14 +1,0 @@
-/* .not-found {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    height: 90vh;
-    text-align: center;
-  }
-  
-  .notFound-text {
-    font-size: 1.5rem;
-    font-family: Asap;
-    padding-bottom: 2rem;
-  } */
